Fix type resolving for thunks.
